+++ variantlib/commands/analyze_platform.py
@@ -7,6 +7,8 @@ import sys
 from variantlib import __package_name__
 from variantlib.commands.plugin_arguments import add_plugin_arguments
 from variantlib.commands.plugin_arguments import parse_plugin_arguments
+from variantlib.plugins.loader import CLIPluginLoader
+from variantlib.plugins.py_envs import ExternalNonIsolatedPythonEnv
 
 logger = logging.getLogger(__name__)
 
@@ -20,15 +22,17 @@ def analyze_platform(args: list[str]) -> None:
     add_plugin_arguments(parser)
 
     parsed_args = parser.parse_args(args)
-    loader = parse_plugin_arguments(parsed_args)
+    plugin_apis = parse_plugin_arguments(parsed_args)
 
-    logger.info("Analyzing the platform ...\n")
-    variant_cfgs = loader.get_supported_configs().values()
+    with ExternalNonIsolatedPythonEnv() as py_ctx:  # noqa: SIM117
+        with CLIPluginLoader(plugin_apis=plugin_apis, python_ctx=py_ctx) as loader:
+            logger.info("Analyzing the platform ...\n")
+            variant_cfgs = loader.get_supported_configs().values()
 
-    # We have to flush the logger handlers to ensure that all logs are printed
-    for handler in logger.handlers:
-        handler.flush()
+            # We have to flush the logger handlers to ensure that all logs are printed
+            for handler in logger.handlers:
+                handler.flush()
 
-    for variant_cfg in variant_cfgs:
-        for line in variant_cfg.pretty_print().splitlines():
-            sys.stdout.write(f"{line}\n")
+            for variant_cfg in variant_cfgs:
+                for line in variant_cfg.pretty_print().splitlines():
+                    sys.stdout.write(f"{line}\n")
