+++ variantlib/commands/plugins/get_supported_configs.py
@@ -8,7 +8,7 @@ from variantlib import __package_name__
 from variantlib.commands.plugins._display_configs import display_configs
 
 if TYPE_CHECKING:
-    from variantlib.loader import PluginLoader
+    from variantlib.plugins.loader import PluginLoader
 
 logger = logging.getLogger(__name__)
